Refactor & format files
